+++ .cmake-format.py
@@ -1,8 +1,18 @@
-additional_commands = {
-  "pkg_find": {
-    "kwargs": {
-      "PKG": "*"
+with section("parse"):
+  additional_commands = {
+    "pkg_find": {
+      "kwargs": {
+        "PKG": "*"
+      }
+    },
+    "format_and_lint": {
+      "kwargs": {
+        "CC": "*",
+        "CCDEPENDS": "*",
+        "CMAKE": "*",
+        "PY": "*",
+        "JS": "*",
+        "EXCLUDE": "*"
+      }
     }
   }
-}
-